pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]
